Add more exceptions to the list

I would also love #667 to be merged, but this probably has to happen after this PR as this PR fixes clippy.

+++ src/report/html.rs
@@ -123,7 +123,7 @@ fn write_report<W: ReportWriter>(
                         result_names.len() - 1
                     });
                 runs[pos] = Some(BuildTestResultHTML {
-                    res: *idx as usize,
+                    res: *idx,
                     log: run.log.clone(),
                 });
             }
